requirements.txt
phidata
openai 
streamlit 
python-dotenv 
requests